requirements.txt
binwalk==2.1.0
cffi==1.14.0
cryptography==2.8
gitdb==4.0.2
GitPython==3.1.0
pycparser==2.19
pygit==0.1
six==1.14.0
smmap==3.0.1
